refactor: unify Program and ProgramV2 into single Program implementation
Remove ProgramV2 as a separate struct and consolidate all program logic
into a single Program that uses SetupEngine for multi-step setup
orchestration (keys, nonces, signatures).

Main changes:
- Delete src/program/program_v2.rs entirely
- Remove dual-routing (get_program_version) from bitvmx.rs
- Remove ProgramVersion enum and version field from ProgramStatus
- Remove setup_v2/SetupV2 API variants — all setup goes through Program
- Simplify ProgramState: remove SettingUpV2 and SettingUp(payload) variants
- Migrate protocol configs (lock, slot, transfer, dispute) to unified API

Code quality:
- Extract try_send_setup_completed() to deduplicate Monitoring/Ready logic
- Extract engine() helper for SetupEngine access
- Reduce log verbosity in hot paths (load, process_comms_message, tick)
- Add warn! when notify_news() is called with no participant keys
- Remove dead StoreKey enum from program.rs
- Clean trailing whitespace
