+++ src/program/mod.rs
@@ -1,6 +1,5 @@
 pub mod participant;
 pub mod program;
-pub mod program_v2;
 pub mod protocols;
 pub mod setup;
 pub mod state;
